Standardise Airbrake url env var name
https://eaflood.atlassian.net/browse/WC-281

Spotted that each app had its own env var for hold the url of the Errbit instance to send exceptions to, but that they were all the same.

In review we would want to point an environment on mass to a particular Errbit instance, not just individual apps so this change is part of a number of changes across the apps to use the same env var for the Errbit instance url.