Mount renewals journey in back office
https://eaflood.atlassian.net/browse/WC-123

Users should be able to renew via an assisted digital route. The back office should support this.

Now that we have the WasteCarriersEngine, we can mount it in the back office to get the renewals journey.

Once we have the journey working in the app, we can start making any additional changes necessary for the AD process.