Move and update key check to include if key is set
In review we realised we shouldn't be inteferring with the secret key if one has already been set. So now we have amended our logic to be set the dummy key if

- we are in production
- rails was started by rake
- the secret key is blank

Because of this change we have also moved it out to a method to make the code clearer.