Move dashboard to /bo
This follows on from https://github.com/DEFRA/waste-carriers-back-office/pull/48

We need to mount the dashboard at /bo so our routing for different applications can be easily configured.

The temporary registrations list page then becomes the root, as this is useful for local testing but won't be accessible in production.