Display current workflow_state in readable format
Rather than just spit out the state from the DB, we can use our existing I18n text to spit out a more human-friendly string.

The order of priority is the page title attribute, then the heading, and if neither are defined then we fall back to the 'raw' workflow_state.