Display matching convictions for main people
It's possible main people could also be flagged for a manual conviction check, so we shouldn't ignore this.

Restructuring the page to group all the people together seemed like a clearer way to display this. However we also state if there are person matches in the convictions section.

WC-376 will go into more detail about reviewing conviction data.