Change ordering of routes in routes.rb
Based on my (limited) knowledge of rails the ordering of routes matters as this sets the order of priority in the rails router.

When trying out some possible solutions to the previous change, I found I could not get to some of the routes I'd set because they instead were being picked up by the WCR engine. Moving the mount call to the end of routes.rb resolved the issue.

So hopefully the new ordering of the routes essentially means

- calls to `/` go here
- these specific calls to `/bo/#` got here
- any other calls to `/bo/#` should just be passed to the engine