Set redirect after sign in and sign out
With these changes now when a user comes to the app e.g. localhost:8000 they'll be redirected to `/bo`. However if not authenticated they'll then be redirected to the user sign in page.

Once authenticated they'll be redirected to `/bo` rather than just `/` as we know that'll just redirect folks to `/bo` anyway.

Finally when users sign out they'll be redirected to `/bo/users/sign_in` rather than `/`, because again they'll eventually land there anyway.

The approach to doing this was taken from https://github.com/plataformatec/devise/wiki/How-To:-redirect-to-a-specific-page-on-successful-sign_in,-sign_out,-and-or-sign_up. There is an alternative approach https://github.com/plataformatec/devise/wiki/How-To%3A-Redirect-to-a-specific-page-on-successful-sign-up-(registration) which moves the redirects to a new controller, but as we don't expect to add anything more to `application_controller.rb` (famous last words!) it was not felt necessary at this time.

All these changes are to ensure that all requests to the app include `/bo` in the path so that our **production** can handle them.

**NOTE** The specs are broken in `spec/requests/dashboards_spec.rb` because they now need to take account of the fact authentication is required in the dashboard. However the work to update the tests has already been done by @irisfaraway in https://github.com/DEFRA/waste-carriers-back-office/pull/76 so waiting for that to be merged before completing the work on this PR.