Fix accidentally pointless tests
Our specs for the search results were checking to see if the term appeared on the page to determine whether the correct result was shown. However, the search term will always appear in the search field, regardless of whether there are matches or not. This meant the tests were completely missing a search-braeking bug.

Checking for the presence of the link to the details page for that renewal is a more precise way to check the results.