Display overview of convictions info, link between pages
There are three types of conviction info - declared_convictions, matches for the business and matches for key people. The page now displays a single-sentence overview of the status for each, plus a line at the top stating whether a check is required or not.

The convictions page now includes a 'back' link to the transient registration details, and the conviction check buttons on the details page now link to the convictions page. It also made sense to add a 'back' link from the details page to the dashboard so things are more joined-up and consistent.