Display relevant people without matching convictions
If the user declared that someone had a relevant conviction, but there were no matches for them, we want to know about it.

This commit also adds some details about each person to aid with making the checks. Since we're reusing some of the rows we use in the transient registration page, let's refactor to use a partial instead.