Working forms for accepting or rejecting conviction
These are workin in the sense that submitting them updates the metaData.revoked_reason. The actual approval or rejection doesn't happen yet.

For these forms, I went back to using form objects and a simplified FormsController to manage loading and submitting the forms. I inherited from the engine where possible, but especially for the controller, there's a lot of functionality (eg managing the workflow state) which means inheritence isn't feasible.