Set up TransferPaymentFormsController
We can now send requests to submit the form with the params. We also add the current user's email to the params so we can include that in the payment details.

There's also a rudimentary view, although this is missing fields – however, it's enough to pass the tests for now.

Also tweaked one of the existing tests so that it doesn't fail if the date in the factory is changed.