Fix post-merge issue with authorising forms
After merging in the new role-based authorisation changes, the payment form broke as we hadn't yet added any authorisation.

We will be adding some authorisation for this form - however, just in case we have forms which don't, we should still fix this issue. This change means that we don't attempt to check authorisation if no auth method is passed in to the AdminFormController from the child form controllers.