Back office users can select which type of payment they want to add
https://eaflood.atlassian.net/browse/WC-377

There are 5 different types of payment which can be manually added. The type of payment a user can add also depends on their role.

Rather than sending the user straight into a form, we first ask them to select which type of payment they want to record, and then direct them to the correct form for that type.