Start working on routing based on form answer
We don't have anything to route to yet (waiting on https://github.com/DEFRA/waste-carriers-back-office/pull/97) but we can implement what happens if the form doesn't submit a valid payment method.

Storing a single list of valid payment methods also means we can use this to reduce duplication in the form.