Seed back office users from json file
https://eaflood.atlassian.net/browse/WC-454

It came up in testing today that we are not able to replicate the work flow we hope to offer the NCCC team when it comes to working with the two systems; the admin section of waste-carriers-frontend and waste-carriers-back-office.

On the day of release they need to be able to access both systems using the same login. So assume they are dealing with a user that needs to renew; they first log into the back end and search for the registration. Once found they click the renew link. They will be taken to the back-office where they will be asked to login. However they can use the same username and password as previously used. Once signed in they can continue as previous, and from this point on they can switch between both systems without needing to re-authenticate.

The problem we have at the moment is that the frontend seeds completely different users to the back-office. So we can't replicate this behaviour in local development and testing. This change covers ensuring we seed the same users created in the frontend into the back-office so we can.

It should be noted there is no way we can remove the need to sign into both systems at least once. Also if a user changes their password in only one of them, they'll need to remember two passwords. And when a new starter joins NCCC they will need to create a user record in each. Hence the desire to move to just one back office system.

But in the meantime we can seed our local environments based on the ideal setup we want in production.

In regards how the change was done, because of the increase in number of users to create and to give us a little more flexibility going forward, this change also switches the logic to create users based on data held in a users.json file (much like how [waste-carriers-front-office](https://github.com/DEFRA/waste-carriers-front-office) does it).