Finance_admin users can record missed WorldPay payment
https://eaflood.atlassian.net/browse/WC-377

A user with the agency or agency_with_refund role should be able to record when a WorldPay payment has been received, but wasn't automatically captured by the usual process (eg the payment was pending and not confirmed immediately).

This works in a very similar way to the bank transfer form we've previously added, although permissions for it are different.