Renewals are completed once there are no pending issues
https://eaflood.atlassian.net/browse/WC-302

Renewals will not be automatically completed if there is a pending conviction check or payment has not been recorded. However, once the conviction check has been approved and the balance is zero, the renewal should automatically complete. This should happen once the last pending action is cleared.

Back office users should also be made aware that clearing a pending issue (eg recording a payment which brings the balance to zero) will also complete the renewal.