Renew after conviction approval form submitted
Once the conviction approval form is successfully submitted, attempt to renew if there are no more pending checks.

A lot of our tests relied on checking the content of the transient_registration after submission. If the renewal completes, that transient_registration no longer exists, so we block completion with a stub for most tests.