Redirect to dashboard if no matching transient reg exists
Normally, after updating a payment or conviction check, the app would redirect to the transient_registration page. However, if the renewal has completed, the transient_registration won't exist and so neither will this page.

Instead, if someone tries to view a non-existent transient_registration, just redirect them to the dashboard.