Convert amount to pence only when creating payment
This fixes a bug discovered during testing. If a payment is added but triggers a validation error, the value in the amount field had already been multiplied by 100. This can lead to an accidental stacking effect when a valid form is eventually submitted with a significantly inflated value. Only converting the value to pence before we pass it in to the method which creates the payment avoids this problem.