Invalidate user session cookies after logout
https://eaflood.atlassian.net/browse/WC-446

After logging out, it's possible to repeat requests made while logged-in by reusing the cookie. This is a known issue with Devise, or more specifically with storing the session in the cookie.

Fortunately, we can fix this by adding a session_token field and using this to extend the salt. Then when a user logs out, we modify the session_token in the DB and the cookie can no longer be reused.