Add working rake task to update addresses
This can be run with `bundle exec rake address:update_from_os_places`. It will only touch registrations which have a past_registration, and transient_registrations which have addresses. And then it will only attempt to update addresses which were originally generated through OS Places results. This means we can use the UPRN to retrieve the data that was previously selected by the user, but this time populate the fields in the new, more consistent way.