Update wcr-renewals to ver. with PDF email attach
https://eaflood.atlassian.net/browse/WC-479

The frontend application generates and attachs a PDF version of the registration (often referred to as the 'certificate') to the confirmation email.

Renewals was shipped without this feature, but always with the intention of adding it later. The functionality has been added to the shared [waste-carriers-renewals](https://github.com/DEFRA/waste-carriers-renewals/) project, so this change just covers bumping back office to a version of renewals that includes it.

🎉 WIP 🎉

N.B. Currently the Gemfile points to a branch in renewals rather than master, as it is waiting for the PR to be completed.