Display expiry date of registration in back office
https://eaflood.atlassian.net/browse/WC-515

Currently you only see details of the renewal, and not the original registration in the back office.

We have had reported issues where a renewal appears to be stuck, and the problem is that the registration has expired which prevents the renewal from completing. However, the only way to confirm this is to look up the original registration.

We believe it will be useful to see the expiry date and status of a registration within the back office, so that NCCC advisors can see why a renewal may now be stuck, and to aid with prioritisation.