Set up RegistrationTransfersController to use form object
Use the form object to validate the params submitted to the form.

This is based off of the AdminFormsController and is pretty similar, with the exception of being specific to a single form and using a registration instead of a transient_registration. Both of these are due a larger refactor to reduce duplication - however that's outside the scope of this PR and will be done separately.