Transfer a registration to a new account
https://eaflood.atlassian.net/browse/WC-61

Now that we can transfer a registration to an existing account, we should also be able to transfer it to a new one.

When the registration transfer form is submitted, if there is no account with the email address used, immediately create a new account and transfer the registration.

The new account should then receive an email notifying them of the change and inviting them to set their password.