Refactor tests so service isn't called twice
Previously some specs called `transfer_to_email` before running the spec, and then called it a second time to test what it returns. This doesn't accurately represent how the service would be used as this should only called once, and calling it twice could produce a different result (for example, if the user doesn't exist at first but is created the second time). So instead we call it in a before block but save the result, and then test that result instead of rerunning it.