Refactor spec to call transfer_to_user every test
Initially we tried to reduce duplication by running transfer_to_user as little as possible and then testing that result multiple times. However this results in the tests being less accurate, and makes it harder to test changes before and after the method is called. So instead we call it in each test at the point at which it's required.