Add rake task to anonymise emails
https://eaflood.atlassian.net/browse/WC-525

We often need to test stuff against production data. If when testing we complete a renewal then the service will send an email to the actual user. We could break the link to our email server within the environment under test, however this could lead to unrelated errors occurring during testing because an email failed to send.

So instead we really need someway that having restored production data ready for debugging/testing, we can then anonymise the email addresses so they will never actually get sent to a real inbox.

🎉 WIP 🎉