Implement anonymise rake task
This is a relatively large change, but the intention is to make rake tasks, even if just for fixes more scalable in the future.

Knowing that this would require us iterating through the entire users collection it was built from the outset to use paging. Plus we would have to connect to 3 collections across both the core and users database. And to keep things low level we use just the MongoDb driver methods directly, something we have done in previous rake tasks.

This meant we were creating very similar methods as currently implemented in [PR 43](https://github.com/DEFRA/waste-carriers-back-office/pull/143), a rake task to fix an address issue.

We also spotted a different take on where the methods sat. Our initial implementation of the anonymise email task called methods defined in the rake file, but within the namespace db::anonymise. Those in the address fix rake task were called outside of its namespace. Some checking highlighted that in both cases they were being added to the global namespace for all rake tasks, and our method names could be clashing (see https://kevinjalbert.com/defined_methods-in-rake-tasks-you-re-gonna-have-a-bad-time/).

So in summary this change also encompasses starting to move some of the similar rake task functionality into the `lib`, from where it can be better managed. We followed advice in [How to structure a Ruby library](https://ryanlue.com/posts/2017-03-03-how-to-structure-a-library) for where to put things, hence we now have

- lib
  - db
    - db.rb
    - anonymise_email.rb
  - db.rb

This also solves an issue rubocop was flagging regards the size of the method block when everything was declared within the namespace in the rake task.