Refactor Db::AnonymiseEmail
Having moved the logic out of the rake task into a class, realised that there was no need to be passing the counts and paging hash between methods; we could just update it directly.

Also decided anything to do with displaying output regards the task should be handled in the task hence the move of the puts statements.

Finally had a go at reducing the Cognitive Complexity CodeClimate highlighted.