Add guard for missing transient_registrations
When this is run in production we know that the transient_registrations collection will exist. However when testing locally we were first using our reset scripts, which recreate the databases, and then seed them.

This process however does not result in the transient_registrations collection being created, so immediately running the rake task would throw an error. Prior to this change the only way to prevent it was to manually create the collection in the database.

Rather than rely on the person using this Rake task knowing this requirement before hand, and to simplify local testing of it, this change adds a guard to handle a missing transient_registrations collection.