Make displaying output configurable
In a previous change we said we wanted the task to handle any output to the console and leave the `Db::AnonymiseEmail` to just do the updates.

However we then remembered `anonymise_email` includes a `puts` statement, and without further mnajor refactoring there is no way the task can handle outputting this.

Hence this change at least allows the calling task to decided whether to display the line by line output or not.