Fix blocked actions post grace window
https://eaflood.atlassian.net/browse/WC-512

During testing of the grace window it was spotted that if a renewal was expired, and it matched during conviction checks, then it could not be continued because the button to access convictions approvals did not display.

On double checking we found built into the view the check `ACTIVE?`, i.e. if the renewal was not ACTIVE then it could not be approved.

🎉 WIP 🎉