Add timings to anonymise email rake task
The rake task will take sometime to run, hence it is likely we would leave it running unattended.

For future planning, and to help determine if further optimisation is needed it would be useful to know approximately how long it takes to run.

Hence this tweak to include timings in the output of the rake task

🎉 WIP 🎉