Tweak to output to improve performance
Displaying the details of the changes for each account can be useful for debug purposes, but will slow things down (if only marginally).

So tweaked the task so now if we want full output, we state we want to debug, else by default the anonymiser will print a period for each page (100) of users it processes, so we have some indication that things are progressing.