Fix issue with trans metadat not supp. may_renew?
We found when looking into an issue in the waste_carriers_engine that we had an assumption we could access the registration from the metaData object, irrespective of whether the metadata was on a registration or a transient_registration.

The actual case is that you can only do it if the metaData is on a registration, hence `may_renew?` will fail if called on a metaData linked to a transient registration.

So this change takes account of that to allow us to call `may_renew?` in the view.