Ensure output to STDOUT is immediate in anonymise
Initial runs of the anonymise email task have shown that it takes a very long time to complete. Therefore its something we want to set running, particularly if we are ssh'd in, then log off and come back to later.

We can do this using something like nohup and redirecting output to a log file, but when we do we have found no output appears in the log file until we command quits or finishes.

Some googling later, and the suggestion is this is due to the output not being flushed. The general concensus was that adding `STDOUT.sync = true` above where your output occurs will resolve the issue.