Update anonymise email rake to use bulk write
When running in our pre-production environment this task is taking more than a day to complete. We therefore started looking into performance improvements.

One idea was to take the risk of duplicates and simply replace the domain in all emails with **example.com**. However we then realised its not actually very easy to update a field based on an existing value in the document when using MongoDb. In fact up to version 3.4 it wasn't even possible.

So instead we found [bulk_write](https://docs.mongodb.com/ruby-driver/master/tutorials/ruby-driver-bulk-operations/). At the moment for each user we are calling 3 different queries;

- update all registrations with a matching account email
- update all transient_registrationsnm with a matching account email
- update the user's email

In each case we are sending a request, waiting for it to complete, and then MongoDb is sending back a result.

With bulk_write we create an array of update statements for each collection, and then apply them as one. So for every 100 users we are just firing one update per collection. The aim is that the reduction in network traffic will result in an improvement in performance.