Add a menu to the back office
https://eaflood.atlassian.net/browse/WC-496

As we add new functionality to the back office, providing links in renewal details is not always going to be the most appropriate place to access the functionality from, and in a lot of cases won't make sense. So we need to add a menu.

This will be used to link to upcoming features like user management, improved convictions workflow, etc.