Add new page for user management
https://eaflood.atlassian.net/browse/WC-495

Agency super users and finance super users should be able to manage back office users - for example, adding or removing users, or changing what permissions they have.

As a starting point for this, we need to have a page which provides an overview of user management which super users can access. This page should be in the menu.