Paginate list of users
We have a lot of back office users, so pagination makes this list much more manageable. We use Kaminari to paginate the list. It shares the same config as the dashboard pagination, showing 20 results per page.

Also change the table to be a full-width column so we have space for the pagination controls.