Refactor rake task into class
Moving the existing code into a class in the lib makes it easier to test if we want to, and removes the duplication we had before.

It also means we don't have to pass a collection to each method call, the relevant collection can be accessed directly in the method that needs it.