Handle users with admin and agency accounts
Because we are syncing across from the backend, we have to deal with the fact that they have individuals who have both an admin and agency user account. This is because in the backend system admins can only manage users, not actually do anything with registrations).

So when syncing agency users, if the role found in the back office is an admin role, we leave it as is rather than overwriting it. This is fine because in the back office admins can do pretty much anything hence only one user account needed.