Amend to capture results and have caller output
We have an idea to allow admin users of the back office system to trigger the sync from the back office UI, rather than rely on using the rake task. This is because as a rake task it has to be kicked off by calling something from with the environment, which requires web-ops involvement and therefore an RfC.

Being able to trigger the sync from the UI will give the team direct control of their user management, and not force them to wait for an RfC to be raised and approved.

However intergrating the sync into the UI is for another PR. This change will just facilitate it.