Copy lib script into service and test it
This commit creates a new UserMigrationService task containing the contents of lib/users/sync_users.rb. The script is unmodified, except for the way we set database collections on initialising - this moves in some logic which was outsourced to a separate Db module before (also in lib).

We also add tests to make sure that the service runs as expected in its current form.