Update FinderService and Dashboard ctrl
With the changes made in waste-carriers-renewals to add scopes to the TransientRegistration model, we need to update the finder to use them, rather than pulling everything back and then querying each object in turn.

It also means we can lean into Kaminari proper, and its paging abilities, so we're also only returning 25 records at a time.

Should note this is not the final version of the class, but good enough we can test it with production data to confirm we actually have a performance improvement.