Set Mongoid log level in production to info
This brings the log level for Mongodb inline with what we set for Rails itself. The default is `:debug` which means literally everything is currently being written to the log.

Not only does it make it very noisy, hopefully we can eek a teensy bit of performance out of the fact we are not having to write to file so much for every query to the database.