Set seaech read preference to secondary
Locally and in our vagrant box this will have no meaning. However in production where we use a replica set, we'll have a primary instance and then a number of secondaries.

The default read mode in Mongoid is primary, which means all queries and processing essentially takes place on the primary server. The benefit is as the primary if you write, and then read you will get back the data written.

The downside is a drop in performance as your query is waiting for a connection on what could be a busy server. By setting the read preference to secondary, you're telling Mongoid (and the MongoDb driver underneath), its ok to go to a secondary instance instead, which will be less busy and therefore should respond quicker. The downside is the risk that you are not looking at the very latest data, as the secondaries are updated after the primary.

However in the case of a dashboard search, we feel this is an acceptable risk as it is unlikely that someone will need to access data immediately after a write.